==> Scripts/ITS/03_ITS_dads_samples.Rmd <==
---
title: "ITS DADS Experiment"
output: html_document
date: "2024-05-21"
---

```{r setup, include=TRUE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = "~/DADS")
library(phyloseq); packageVersion("phyloseq")
library(ggplot2); packageVersion("ggplot2")
library(readr); packageVersion("readr")
library(purrr); packageVersion("purrr")
library(furrr); packageVersion("furrr")
library(dplyr); packageVersion("dplyr")
library(stringr); packageVersion("stringr")
library(metacoder); packageVersion("metacoder")
library(data.table); packageVersion("data.table")
library(Biostrings); packageVersion("Biostrings")
library(vegan);packageVersion("vegan")
library(agricolae);packageVersion("agricolae")
library(magrittr);packageVersion("magrittr")
library(data.table);packageVersion("data.table")
library(ggbeeswarm);packageVersion("ggbeeswarm")
library(DESeq2);packageVersion("DESeq2")
library(cowplot);packageVersion("cowplot")
theme_set(theme_cowplot())
library(doParallel);packageVersion("doParallel")
library(viridis);packageVersion("viridis")
```

### Load data
```{r subset data}
load("Data/ITS/filtered_data/dadsexp.RData")
load("Data/ITS/filtered_data/dads_5asv.RData")

filtered_read_counts<-read.csv("Data/ITS/filtered_data/ITS_subset_track_reads_dads.csv")
summary(filtered_read_counts)
```

### Rarefaction curve 
```{r rarefaction curve}
### Rarefaction curve 
otu_data <- as(otu_table(dads_5asv.ps), "matrix")

S <- specnumber(otu_data) # observed number of species
raremax <- min(rowSums(otu_data))

pdf("Figures/ITS/rarefaction_curve.pdf", width = 10, height = 8)
rarecurve(otu_data, step = 10, sample = raremax, col = "blue", cex = 0.6, label = TRUE, xlab = "Sequencing Depth", ylab = "Number of ASVs")
dev.off()

# Save the plot to a PDF
pdf("Figures/ITS/speciesaccum_curve.pdf", width = 10, height = 8)
asv_accum <- vegan::specaccum(otu_data, method = "exact")
plot(asv_accum, ci.type = "poly", col = "blue", lwd = 2, ci.lty = 0, ci.col = "lightblue", xlab = "Number of samples", ylab = "Number of Filtered ASVs")
dev.off()
```
```

### Bar plot agglomerated by species-using filtered dataset and facet wrapped by treatment
```{r barplot-Relative abundance-treatment}
ITSdata <- dads_5asv.ps %>%
  tax_glom(taxrank = "Genus") %>%                     # agglomerate at Genus level
  transform_sample_counts(function(x) {x/sum(x)} ) %>% # Transform to rel. abundance
  psmelt() %>%                                         # Melt to long format
  #filter(Abundance > 0.02) %>%                       # Filter out low abundance taxa
  arrange(desc(Abundance))
                                # Sort data frame alphabetically by phylum

# Plot 
abund_plot <- ggplot(ITSdata, aes(x = Sample, y = Abundance, fill = Genus)) + 
  geom_bar(stat = "identity", position = "stack", color = "black", size = 0.2) +
  scale_fill_viridis_d() +
  scale_color_viridis_d() +
  theme_minimal() +
  labs(
    y = "Relative Abundance",
    fill = "Species"
  ) +
  facet_grid(~factor(Treatment), scales = "free_x") +
  theme(
    axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5, size = 14),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    legend.position = "bottom",
    legend.text = element_text(size = 14),
    legend.title = element_text(size = 20),  # Adjust legend title size
    legend.key.size = unit(2, "lines"),  # Adjust the size of the legend color key
    strip.text = element_text(size = 20),
    strip.background = element_blank()
  ) +
  guides(
    fill = guide_legend(
      reverse = FALSE,
      keywidth = 1,
      keyheight = 1,
      title.position = "top",
      title.hjust = 3,  # Center the legend title
      label.theme = element_text(size = 18)  # Adjust the size of the legend labels
    )
  )

abund_plot

ggsave("Figures/ITS/relabundance_plot_treatment_relabund_genus_dads.pdf", abund_plot, width = 22, height = 16)
```

### Bar plot agglomerated by species-using filtered dataset and faceted by soil 
```{r barplot-Relative abundance-soil}
ITSdata <- dads_5asv.ps %>%
  tax_glom(taxrank = "Genus") %>%                     # agglomerate at Genus level
  transform_sample_counts(function(x) {x/sum(x)} ) %>% # Transform to rel. abundance
  psmelt() %>%                                         # Melt to long format
  #filter(Abundance > 0.02) %>%                       # Filter out low abundance taxa
  arrange(desc(Abundance))
                                # Sort data frame alphabetically by phylum

# Plot 
abund_plot <- ggplot(ITSdata, aes(x = Sample, y = Abundance, fill = Genus)) + 
  geom_bar(stat = "identity", position = "stack", color = "black", size = 0.2) +
  scale_fill_viridis_d() +
  scale_color_viridis_d() +
  theme_minimal() +
  labs(
    y = "Relative Abundance",
    fill = "Species"
  ) +
  facet_grid(~factor(Soil), scales = "free_x") +
  theme(
    axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5, size = 14),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    legend.position = "bottom",
    legend.text = element_text(size = 14),
    legend.title = element_text(size = 20),  # Adjust legend title size
    legend.key.size = unit(2, "lines"),  # Adjust the size of the legend color key
    strip.text = element_text(size = 20),
    strip.background = element_blank()
  ) +
  guides(
    fill = guide_legend(
      reverse = FALSE,
      keywidth = 1,
      keyheight = 1,
      title.position = "top",
      title.hjust = 3,  # Center the legend title
      label.theme = element_text(size = 18)  # Adjust the size of the legend labels
    )
  )

abund_plot

ggsave("Figures/ITS/relabundance_plot_soil_relabund_genus_dads.pdf", abund_plot, width = 22, height = 16)
```

### Bar plot agglomerated by species-using filtered dataset and faceted by incubation
```{r barplot-Relative abundance-incubation}
ITSdata <- dads_5asv.ps %>%
  tax_glom(taxrank = "Species") %>%                     # agglomerate at Genus level
  transform_sample_counts(function(x) {x/sum(x)} ) %>% # Transform to rel. abundance
  psmelt() %>%                                         # Melt to long format
  #filter(Abundance > 0.02) %>%                       # Filter out low abundance taxa
  arrange(desc(Abundance))
                                # Sort data frame alphabetically by phylum

# Plot 
abund_plot <- ggplot(ITSdata, aes(x = Sample, y = Abundance, fill = Genus)) + 
  geom_bar(stat = "identity", position = "stack", color = "black", size = 0.2) +
  scale_fill_viridis_d() +
  scale_color_viridis_d() +
  theme_minimal() +
  labs(
    y = "Relative Abundance",
    fill = "Genus"
  ) +
  facet_grid(~factor(Incubation, labels = c("Anaerobic", "Aerobic")), scales = "free_x") +
  theme(
    axis.text.x = element_text(angle = 90, hjust = 1, vjust = 0.5, size = 14),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    legend.position = "bottom",
    legend.text = element_text(size = 14),
    legend.title = element_text(size = 20),  # Adjust legend title size
    legend.key.size = unit(2, "lines"),  # Adjust the size of the legend color key
    strip.text = element_text(size = 20),
    strip.background = element_blank()
  ) +
  guides(
    fill = guide_legend(
      reverse = FALSE,
      keywidth = 1,
      keyheight = 1,
      title.position = "top",
      title.hjust = 3,  # Center the legend title
      label.theme = element_text(size = 18)  # Adjust the size of the legend labels
    )
  )

abund_plot

ggsave("Figures/ITS/relabundance_plot_incubation_relabund_genus_dads.pdf", abund_plot, width = 22, height = 16)
```

### Let's look top taxa 
```{r Top taxa}
new_names <- sapply(strsplit(dads_5asv.ps@sam_data$Sample_nameITS, "-"), `[`, 1)

# Rename the Sample_nameITS column
dads_5asv.ps@sam_data$Sample_nameITS <- new_names

top40 <- names(sort(taxa_sums(dads_5asv.ps), decreasing=TRUE))[1:40]
ps.top40 <- transform_sample_counts(dads_5asv.ps, function(OTU) OTU/sum(OTU))
ps.top40<- tax_glom(ps.top40, taxrank = "Species") 
ps.top40 <- prune_taxa(top40, ps.top40)

pdf("Figures/ITS/top40_species_dads.pdf", width = 34, height = 20)
plot_bar(ps.top40, x="Sample_nameITS", fill="Species") + facet_wrap(~Treatment, scales="free_x")
dev.off()

plot_bar(ps.top40, x="Sample_nameITS", fill="Species") + facet_wrap(~Treatment, scales="free_x")

data <-
  dads_5asv.ps %>%
  tax_glom("Species") %>%
  psmelt() %>%
  as_tibble()

data_summed <- data %>%
  filter(Abundance > 0) %>%
  group_by(Kingdom, Phylum, Class, Order, Family, Genus, Species) %>%
  summarise(Abundance = sum(Abundance),
            Prevalence = n_distinct(Sample)) %>%
  ungroup() %>%
  mutate(Total_Abundance = sum(Abundance)) %>%
  mutate(Percentage = 100*(Abundance / Total_Abundance))
  
write.csv(data_summed, file="Results/ITS/mostabundanttaxa_abundance_dads_species.csv",row.names = TRUE)

data <-
  dads_5asv.ps %>%
  tax_glom("Genus") %>%
  psmelt() %>%
  as_tibble()

data_summed <- data %>%
  filter(Abundance > 0) %>%
  group_by(Kingdom, Phylum, Class, Order, Family, Genus) %>%
  summarise(Abundance = sum(Abundance),
            Prevalence = n_distinct(Sample)) %>%
  ungroup() %>%
  mutate(Total_Abundance = sum(Abundance)) %>%
  mutate(Percentage = 100*(Abundance / Total_Abundance))
  
write.csv(data_summed, file="Results/ITS/mostabundanttaxa_abundance_dads_genus.csv",row.names = TRUE)

data1 <-
  dads_5asv.ps %>%
  tax_glom("Species") %>%
  transform_sample_counts(function(x) {x/sum(x)} ) %>%
  psmelt() %>%
  as_tibble()

data_summed1 <- data1 %>%
  filter(Abundance > 0) %>%
  group_by(Treatment, Kingdom, Phylum, Class, Order, Family, Genus, Species) %>%
  summarise(Abundance = sum(Abundance),
            Prevalence = n_distinct(Sample)) 

write.csv(data_summed1, file="Results/ITS/mostabundanttaxa_relativeabundance_untreatedtreated_dads.csv",row.names = TRUE)

data2 <-
  dads_5asv.ps %>%
  tax_glom("Species") %>%
  psmelt() %>%
  as_tibble()

# Calculate percentages
data_summed2 <- data2 %>%
  filter(Abundance > 0) %>%
  group_by(Treatment, Kingdom, Phylum, Class, Order, Family, Genus, Species) %>%
  summarise(Abundance = sum(Abundance),
            Prevalence = n_distinct(Sample))

write.csv(data_summed2, file="Results/ITS/mostabundanttaxa_untreatedtreated_dads.csv",row.names = TRUE)
```
### Which samples of Stromatinia in them? 
```{r Identify samples with stromatinia}
tax_table <- as.data.frame(dads.ps@tax_table)

stromatinia_taxa <- apply(tax_table(dads.ps), 1, function(x) any(grepl("Stromatinia", x, ignore.case = TRUE)))
print(sum(stromatinia_taxa))

stromatinia_asvs <- rownames(tax_table(dads.ps))[stromatinia_taxa]

otu_table <- as.data.frame(as(otu_table(dads.ps), "matrix"))

if (ncol(otu_table) == nsamples(dads.ps)) {
  # If samples are in columns, transpose the table
  otu_table <- t(otu_table)
}

if ("ASV784" %in% colnames(otu_table)) {
  stromatinia_counts <- otu_table[, "ASV784", drop = FALSE]
  
  # Create a data frame with sample names and Stromatinia counts
  result <- data.frame(
    Sample = rownames(stromatinia_counts),
    Stromatinia_Count = stromatinia_counts[,1]
  )
  
  # Remove any samples with zero counts (if desired)
  result <- result[result$Stromatinia_Count > 0, ]
  
  # Print the dimensions and first few rows of the result
  print(dim(result))
  print(head(result))
} else {
  print("ASV784 not found in the OTU table.")
  
  # Print the first few column names of the OTU table for verification
  print("First few ASVs in the OTU table:")
  print(head(colnames(otu_table)))
}

```

### Let's normalize our data in preparation for beta diversity analyses
```{r Normalization}
#Prune ASVs not present in least 5 times (across one more more samples)
#dads_5asv.ps<- phyloseq::prune_taxa(taxa_sums(dads.ps) > 5, dads.ps)
sdt = data.table::data.table(as(sample_data(dads_5asv.ps), "data.frame"),
                 TotalReads = sample_sums(dads_5asv.ps), keep.rownames = TRUE)
min_reads<-min(sdt$TotalReads)

rarefied.ps <- rarefy_even_depth(dads_5asv.ps, sample.size = min_reads, rngseed = 100)

#Relative abundance
strat_rel.ps<-transform_sample_counts(dads_5asv.ps, function(x) x / sum(x))
```

### Alpha diversity-use non-normalized or pruned data
Let's start with t test as point of comparison with anova analysis
```{r alpha diversity analyses-t-test}
set.seed(1)
AD <- estimate_richness(dads.ps, measures = c("Observed", "InvSimpson", "Shannon"))
rownames(AD) <- gsub("\\.", "-", rownames(AD))

# Combine richness estimates with sample metadata
AD <- cbind(AD, sample_data(dads.ps)[row.names(AD), c("Treatment", "Incubation", "Soil")])

# Convert to data.table and keep row names as a column
AD <- data.table(AD, keep.rownames = TRUE)

# Reshape the data frame
AD <- melt(AD, id.vars = c('rn', c("Treatment", "Incubation", "Soil")))

# Now conduct t-tests (assuming you have appropriate factors)
factors <- c("Treatment", "Incubation", "Soil")
# Iterate through each factor
ttest_list <- list()

# Iterate through each factor
for (factor_name in factors) {
  # Initialize an empty data.table to store t-test results for each alpha diversity measure
  ttest <- data.table(alpha_index = character(), pvalue = numeric())
  
  # Iterate through each alpha diversity measure
  for (alphad in c("InvSimpson", "Shannon", "Observed")) {
    # Perform t-test
    ttest_res <- t.test(value ~ get(factor_name), data = AD[variable == alphad], var.equal = TRUE)
    
    # Extract p-value from t-test results and store it in the data.table
    p_value <- ttest_res$p.value
    ttest <- rbind(ttest, data.table(alpha_index = alphad, pvalue = p_value))
  }
  
  # Add t-test results for the current factor to the list
  ttest_list[[factor_name]] <- ttest
}

# Print t-test results
for (factor_name in factors) {
  cat(paste("T-Test Results for Factor:", factor_name), "\n")
  print(knitr::kable(ttest_list[[factor_name]], format = "markdown"))
  cat("\n")
}
``` 


### Let's no proceed with ANOVA analyses and then compile our alpha diversity stats
```{r compiled alpha diversity table}
set.seed(1)
# Initialize an empty data.table to store combined ANOVA results for all factors
anova_list <- list()

factors2 <- c("Treatment", "Incubation", "Soil")

# Initialize a vector to store factor names for which a plot has been created
plotted_factors <- character()
 

final_anova_table <- data.table(factor_name = character(), alpha_index = character(), pvalue = numeric(), df_between = numeric(), df_within = numeric(), ss_between = numeric(), ss_within = numeric(), ms_between = numeric(), ms_within = numeric(), F_value = numeric())

for (factor_name in factors2) {
  # Initialize an empty data.table to store ANOVA results for each alpha diversity measure
  anova <- data.table(factor_name = character(), alpha_index = character(), pvalue = numeric(), df_between = numeric(), df_within = numeric(), ss_between = numeric(), ss_within = numeric(), ms_between = numeric(), ms_within = numeric(), F_value = numeric())
  
  # Iterate through each alpha diversity measure
  for (alphad in c("Observed","InvSimpson", "Shannon")) {
    # Perform ANOVA
    anova_res <- aov(value ~ get(factor_name), data = AD[variable == alphad])
    
    # Extract relevant statistics from ANOVA results
    anova_summary <- summary(anova_res)
    
    # Extract p-value from ANOVA results
    p_value <- anova_summary[[1]]$`Pr(>F)`[1]
    
    # Extract other statistics
    df_between <- anova_summary[[1]]$"Df"[1]
    df_within <- anova_summary[[1]]$"Df"[2]
    ss_between <- anova_summary[[1]]$"Sum Sq"[1]
    ss_within <- anova_summary[[1]]$"Sum Sq"[2]
    ms_between <- anova_summary[[1]]$"Mean Sq"[1]
    ms_within <- anova_summary[[1]]$"Mean Sq"[2]
    F_value <- anova_summary[[1]]$"F value"[1]
    
    # Store statistics in the data.table
    anova <- rbind(anova, data.table(factor_name = factor_name, alpha_index = alphad, pvalue = p_value, df_between = df_between, df_within = df_within, ss_between = ss_between, ss_within = ss_within, ms_between = ms_between, ms_within = ms_within, F_value = F_value))
  }
  
  # Append ANOVA results for the current factor to the final table
  final_anova_table <- rbind(final_anova_table, anova)
}

# Print final ANOVA table
print(knitr::kable(final_anova_table, format = "markdown"))
write.csv(final_anova_table, "Results/ITS/final_anova_table_dads.csv", row.names = FALSE)
```

### Let's make faceted boxplots for our alpha diversity analyses
```{r alpha diversity faceted plots}
set.seed(1)
# Estimate alpha diversity
AD <- estimate_richness(dads.ps, measures = c("Observed", "InvSimpson", "Shannon"))
rownames(AD) <- gsub("\\.", "-", rownames(AD))

# Combine richness estimates with sample metadata
AD <- cbind(AD, sample_data(dads_5asv.ps)[rownames(AD), c("Treatment", "Soil", "Incubation")])


variables <- c("Treatment", "Soil", "Incubation")


# Create an empty list to store plots
plots <- list()

# Loop through each variable and create the plot
for (variable in variables) {
  p <- plot_richness(dads.ps, x = variable, measures = c("Shannon", "InvSimpson", "Observed")) + 
    geom_boxplot()
  plots[[variable]] <- p
}

# Arrange plots in a grid
plots_arrange <- gridExtra::grid.arrange(
  plots[["Treatment"]], plots[["Incubation"]], plots[["Soil"]])
  nrow = length(variables)


# Add annotations to the top of each plot
plots_arrange_with_text <- cowplot::ggdraw() +
  cowplot::draw_plot(plots_arrange) +
  cowplot::draw_plot("Treatment", x = 0.28, y = 0.98) +
  cowplot::draw_plot("Incubation", x = 0.73, y = 0.98) +
  cowplot::draw_plot("Soil", x = 0.73, y = 0.98)

# Save the arranged plots
ggsave(plots_arrange_with_text, path = "Figures/ITS", filename = "alpha_diversity_each_factor_dads.pdf", 
       width = 12, height = 16)
```

### Let's perform TUKEY HSD tests
```{r alpha diversity compilation with tukey summary}
set.seed(1)
# Define your variables of interest
# Define your variables of interest
variables <- c("Treatment", "Soil", "Incubation")
measures <- c("Observed", "InvSimpson", "Shannon")

# Create an empty list to store ANOVA and Tukey's results
tukey_results <- list()

# Loop through each variable and measure
for (variable in variables) {
  for (measure in measures) {
    # Subset the data for the current variable and measure
    data_subset <- AD[, c(variable, measure)]
    
    # Perform ANOVA
    formula <- as.formula(paste(measure, "~ ."))
    anova_res <- aov(formula, data = data_subset)
    
    # Perform Tukey's post-hoc test
    tukey_result <- HSD.test(anova_res, variable)
    
    # Store Tukey's results
    print(tukey_result)
  }
}
```

### Beta diversity-PERMANOVA analysis on rarefied read counts
```{r PERMANOVA ANALYSES-rarefied}
set.seed(1)
#I used the filtered dataset (removed low abundance counts less than 2)
mapping <- as.data.frame(sample_data(rarefied.ps))

mapping2=mapping[, c("Treatment", "Soil", "Incubation")]

ASV_tables_bray_rarefy <- phyloseq::distance(rarefied.ps, method = "bray")

df_metadata <- data.frame(sample_data(rarefied.ps))

df_metadata<-df_metadata[, c("Treatment", "Soil", "Incubation")]

pvals <- sapply(colnames(mapping2), function(variable){
  
  #Create formula for adonis() 
  form1<-as.formula(paste("ASV_tables_bray_rarefy", variable, sep="~"))
  
  res <- adonis2(form1, data = df_metadata, permutations = 9999)
  
  # extract pvalue from adonis output
  pval <- res$`Pr(>F)`[[1]] # we take the first p value which corresponds to the variable of interest
  return(pval)
})

#Adjust p_values with fdr correction
pval_adj <- p.adjust(pvals, method = "fdr")

#Find significant variables with lower than 0.05 for their p values
save.pval.adonis0.05_braycurtis <- pval_adj[pval_adj < 0.05]
save.pval.adonis0.05_braycurtis
print(pval_adj)
```

### Beta diversity-PERMANOVA analyses-ASV counts transformed (rel. abundance)
```{r PERMANOVA ANALYSES-relative abundance, include=TRUE}
set.seed(1)
#I used the filtered dataset (removed low abundance counts less than 2)
mapping <- as.data.frame(sample_data(strat_rel.ps))

mapping2=mapping[, c("Treatment", "Soil", "Incubation")]

ASV_tables_bray <- phyloseq::distance(strat_rel.ps, method = "bray")

df_metadata <- data.frame(sample_data(strat_rel.ps))

df_metadata<-df_metadata[, c("Treatment", "Soil", "Incubation")]

pvals <- sapply(colnames(mapping2), function(variable){
  
  #Create formula for adonis() 
  form1<-as.formula(paste("ASV_tables_bray", variable, sep="~"))
  
  res <- adonis2(form1, data = df_metadata, permutations = 9999)
  
  # extract pvalue from adonis output
  pval <- res$`Pr(>F)`[[1]] # we take the first p value which corresponds to the variable of interest
  return(pval)
})

#Adjust p_values with fdr correction
pval_adj <- p.adjust(pvals, method = "fdr")

#Find significant variables with lower than 0.05 for their p values
save.pval.adonis0.05_braycurtis <- pval_adj[pval_adj < 0.05]
save.pval.adonis0.05_braycurtis

print(pval_adj)
```


### Compile stats
```{r compile the stats}
set.seed(1)
# Perform PERMANOVA with interactions
res <- adonis2(ASV_tables_bray ~ Treatment*Incubation*Soil, data = df_metadata, permutations = 9999, by="terms")

# Extract data from res
res_data <- data.frame(res)
colnames(res_data)[colnames(res_data) == "Pr..F."] <- "pr_f"

# Remove the last two rows (Residual and Total)
res_data <- res_data[-c(nrow(res_data)-1, nrow(res_data)),]

# Add FDR and Bonferroni adjusted p-values
res_data$FDR_adj_p <- p.adjust(res_data$pr_f, method = "fdr")
res_data$Bonferroni_adj_p <- p.adjust(res_data$pr_f, method = "bonferroni")
res_data$Factor <- rownames(res_data)

# Reorder columns
res_data <- res_data[, c("Factor", "Df", "SumOfSqs", "R2", "F", "pr_f", "FDR_adj_p", "Bonferroni_adj_p")]
rownames(res_data) <- NULL

# Perform beta dispersion analysis for main effects
beta_disp_results <- lapply(c("Treatment", "Incubation", "Soil"), function(variable) {
  res_betadisper <- betadisper(d = ASV_tables_bray, group = df_metadata[[variable]])
  perm_test_res <- permutest(res_betadisper)
  pval_betadisper <- perm_test_res$tab$`Pr(>F)`[[1]]
  
  data.frame(Factor = variable, P_Value_BetaDispersion = pval_betadisper)
})

beta_disp_results_df <- do.call(rbind, beta_disp_results)

# Merge PERMANOVA results with beta dispersion results
final_results <- merge(res_data, beta_disp_results_df, by = "Factor", all.x = TRUE)

# Reorder columns
final_results <- final_results[, c("Factor", "Df", "SumOfSqs", "R2", "F", "pr_f", "FDR_adj_p", "Bonferroni_adj_p", "P_Value_BetaDispersion")]

# Print the table
print(final_results, row.names = FALSE)

# Save results
write.csv(final_results, file = "Results/ITS/PERMANOVA_interactions_with_betadisper.csv", row.names = FALSE)
```

### What about interactions between factors?
```{r pairwise permanova comparisons}
set.seed(1)
# Calculate PERMANOVA
ASV_tables_bray <- phyloseq::distance(strat_rel.ps, method = "bray")
df_metadata <- data.frame(sample_data(strat_rel.ps))
df_metadata <- df_metadata[, c("Treatment", "Soil", "Incubation")]

# Define function to perform PERMANOVA for pairwise combinations of factors
perform_permutation <- function(factor1, factor2) {
  interaction <- paste(factor1, factor2, sep = " * ")
  res <- adonis2(ASV_tables_bray ~ ., data = df_metadata[, c(factor1, factor2)], permutations = 9999)
  return(list(interaction = interaction, p_value = res$`Pr(>F)`[[1]]))
}

# Perform pairwise comparisons between factors
significant_interactions <- sapply(names(df_metadata), function(factor1) {
  sapply(names(df_metadata), function(factor2) {
    if (factor1 != factor2) {
      perform_permutation(factor1, factor2)
    } else {
      NULL
    }
  })
}, simplify = "array")

# Filter out NULLs and combine results
significant_interactions <- significant_interactions[!sapply(significant_interactions, is.null)]

# Print significant interactions
print(significant_interactions)

interactions_df <- do.call(rbind.data.frame, significant_interactions)
print(significant_interactions)

interactions_df <- do.call(rbind.data.frame, significant_interactions)


adjusted_p_values <- p.adjust(interactions_df$p_value, method = "fdr")
# Add adjusted p-values to the data frame
interactions_df$p_value <- adjusted_p_values


# Extract unique factors
unique_factors <- unique(unlist(strsplit(interactions_df$interaction, " \\* ")))

# Create an empty matrix to store p-values
interaction_matrix <- matrix(NA, nrow = length(unique_factors), ncol = length(unique_factors), dimnames = list(unique_factors, unique_factors))

# Fill the matrix with p-values
for (i in 1:nrow(interactions_df)) {
  factors <- unlist(strsplit(interactions_df$interaction[i], " \\* "))
  interaction_matrix[factors[1], factors[2]] <- interactions_df$p_value[i]
  interaction_matrix[factors[2], factors[1]] <- interactions_df$p_value[i]
}

rownames(interaction_matrix) <- unique_factors 
colnames(interaction_matrix) <- unique_factors 


# Convert the matrix to a data frame for ggplot
interaction_df <- as.data.frame(as.table(interaction_matrix))
interaction_df <- as.data.frame(as.table(interaction_matrix))
interaction_df$color <- ifelse(is.na(interaction_df$Freq), "gray",
                               ifelse(interaction_df$Freq <= 0.01, "darkblue",
                                      ifelse(interaction_df$Freq <= 0.05, "lightblue", "white")))

# Plot heatmap with continuous color scale
heatmap <- ggplot(data = interaction_df, aes(x = Var1, y = Var2, fill = color)) +
  geom_tile(color = "black") +  # Add border for each tile
  scale_fill_manual(values = c("gray", "darkblue", "white"),
                    name = "P-value",
                    labels = c("NA", "P < 0.01", "P > 0.05"),
                    na.value = "gray") +  # Set manual color scale and legend labels
  scale_fill_identity(name = NULL) +  # Use identity scale for custom colors
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1, size = 10),  # Adjust x-axis text angle and size
        axis.text.y = element_text(size = 10),  # Adjust y-axis text size
        panel.grid = element_blank(),  # Remove gridlines
        panel.border = element_blank(),  # Remove border around the plot
        axis.line = element_blank(),  # Remove axis lines
        legend.position = "bottom",  # Position legend at the bottom
        legend.direction = "horizontal",  # Horizontal legend
        legend.key.size = unit(0.5, "cm"),  # Adjust legend key size
        legend.title = element_blank(),  # Remove legend title
        legend.text = element_text(size = 8))  # Adjust legend text font size

# Assign row and column names
heatmap <- heatmap + labs(x = NULL, y = NULL)

# Print heatmap
print(heatmap)

# Save the heatmap as a PDF file
ggsave("Figures/ITS/permanova_interactions_heatmap_dads.pdf", plot = heatmap, width = 8, height = 6)
```

### Let's now examine beta diversity through ordination plots
We'll will examine outputs for rarefied data as well as data adjusted by relative abundance
We'll now make ordination plots on data for which we've calculated relative abundance
```{r Beta diversity-Rel. abundance-treated and untreated samples}
set.seed(1)
#Relative abundance
ps_PCoA3 = ordinate(strat_rel.ps, 
                   method="PCoA", 
                   distance="bray")

plot_3<- plot_ordination(strat_rel.ps, 
                         ps_PCoA3,
                         color= "Treatment") +
  ggtitle("PCoA ordination plot - relative abundance") +
  labs(color = "Factor") + 
  theme(axis.text=element_text(size=14), axis.title=element_text(size=14), title =  element_text(size=13), legend.text = element_text(size = 18))

plot_ellipses<-plot_3 +
   stat_ellipse(type = "norm", linetype = 2) +
  theme_bw()

plot_ellipses

#Export PCoA coordinates per sample
write.csv(ps_PCoA3[["points"]], file="Results/ITS/ITS_PCoA_coordinates_persample.csv",row.names = TRUE)


res <- adonis2(ASV_tables_bray~Treatment*Incubation*Soil, data = df_metadata, permutations = 9999)
res
pval=res$`Pr(>F)`[2]
pval_adj <- p.adjust(pval, method = "fdr")
print(pval_adj)
```

Now, we'll look at ordination plot (rel abundance dataset), with incubation factor
```{r Beta diversity-Rel. abundance-incubation}
set.seed(1)
#Relative abundance
ps_PCoA3 = ordinate(strat_rel.ps, 
                   method="PCoA",
                   distance="bray")

plot_4<- plot_ordination(strat_rel.ps, 
                         ps_PCoA3,
                         labs(color = "Factor"),
                         color= "Incubation") +
  ggtitle("PCoA ordination plot - relative abundance") +
  theme(axis.text=element_text(size=14), axis.title=element_text(size=14), title =  element_text(size=13), legend.text = element_text(size = 18))


plot_ellipses<-plot_4 +
   stat_ellipse(type = "norm", linetype = 2) +
  theme_bw()

plot_ellipses
```

Now, we'll look at ordination plot rel abundance-with Soil factor
```{r Beta diversity-Rel. abundance-soil}
set.seed(1)
#Relative abundance
ps_PCoA3 = ordinate(strat_rel.ps, 
                   method="PCoA",
                   distance="bray")

plot_4<- plot_ordination(strat_rel.ps, 
                         ps_PCoA3,
                         labs(color = "Factor"),
                         color= "Soil") +
  ggtitle("PCoA ordination plot - relative abundance") +
  theme(axis.text=element_text(size=14), axis.title=element_text(size=14), title =  element_text(size=13), legend.text = element_text(size = 18))


plot_ellipses<-plot_4 +
   stat_ellipse(type = "norm", linetype = 2) +
  theme_bw()

plot_ellipses
```

### Functions to prepare to do differential abundance analyses
```{r organizing data for downstream steps}
sample.data.table <- function(ps) {
  as(sample_data(ps), "data.frame") %>%
    as.data.table(keep.rownames = "Sample") %>% 
    setkey(Sample) %>%
    return()
}
asv.data.table <- function(ps) {
  if (taxa_are_rows(ps)) {
    otu.mat <- t(as(otu_table(ps), "matrix"))
  } else {
    otu.mat <- as(otu_table(ps), "matrix")
  }
  return(as.data.table(otu.mat, keep.rownames = "Sample") %>% setkey(Sample))
}
cat.n <- function(x) cat(x, sep = "\n")
```

### Now we import data
```{r data-import-and-prep}
dads.data <- sample.data.table(dads_5asv.ps)
dads.asvs <- asv.data.table(dads_5asv.ps)
dads.asv.ids <- taxa_names(dads_5asv.ps)
dads.dt <- dads.data[dads.asvs]

#Let's adjust treatment factor so order is in line with metacoder outputs
dads_5asv.ps@sam_data$Treatment <- factor(dads_5asv.ps@sam_data$Treatment,
                                     levels = c("DADS_treated","untreated"))
```

### Conduct Wilcoxon-rank sum tests on the dads data set-treatment
```{r wilcoxon-tests-treatuntreat}
set.seed(1)
wilcox.tests <- lapply(dads.asv.ids, function(asv) {
  frm <- paste(asv, "~ Treatment") # this combined with the as.formula() function below makes it possible to generate formulas in R programmatically
  return(
    wilcox.test(as.formula(frm), data = dads.dt)
  )
})

dads.raw.pvals <- sapply(wilcox.tests, function(mod) {return(mod$p.value)})

dads.fdr.pvals <- p.adjust(dads.raw.pvals, method = "fdr")

paste("Significant associations:", length(dads.fdr.pvals[dads.fdr.pvals <= 0.05])) %>%
  cat.n()

dads.raw.pvals <- sapply(wilcox.tests, function(mod) mod$p.value)
dads.fdr.pvals <- p.adjust(dads.raw.pvals, method = "fdr")

# Create data table of significant results
dads.wilcox.dt <- data.table(
  ASV = dads.asv.ids,
  P_Value = dads.raw.pvals,
  Adjusted_P_Value = dads.fdr.pvals
)

dads.wilcox.dt <- dads.wilcox.dt[order(Adjusted_P_Value)]

# Filter rows where Adjusted_P_Value < 0.05
dads.wilcox.dt <- dads.wilcox.dt[Adjusted_P_Value < 0.05]

# Optionally, merge taxonomic information if available
tax_table <- as.data.frame(tax_table(dads_5asv.ps))
tax_table$ASV_id <- rownames(tax_table)
dads.wilcox.dt <- merge(dads.wilcox.dt, tax_table, by.x = "ASV", by.y = "ASV_id", all.x = TRUE)

# Export to CSV
write.csv(dads.wilcox.dt, file = "Results/ITS/wilcox_treat_vs_untreat_dads.csv", row.names = FALSE)
```

### Let's now do a DESEQ2 analysis-treatment
```{r deseq2-trt}
set.seed(1)
dads.deseq <- phyloseq_to_deseq2(dads_5asv.ps, ~ Treatment) # this function does all the heavy lifting for us to turn a phyloseq object into a DESeq2 object

#This was required so DESEQ doesn't error out (there are several approaches, but the deseq developer recommended this most recently, if going from phyloseq to deseq option)
dads.dds <- estimateSizeFactors(dads.deseq, type = 'poscounts')
dads.dds = DESeq(dads.dds)
dads.res.dt <- results(dads.dds) %>% as.data.table(keep.rownames = "ASV")

dads.sig.dt <- dads.res.dt[padj <= 0.05]

dads.sig.dt = cbind(as(dads.sig.dt, "data.frame"), as(tax_table(dads_5asv.ps)[dads.sig.dt$ASV, ], "matrix"))

dads.sig.dt <- dads.sig.dt %>%
  mutate(across(everything(), ~ifelse(is.na(.), "Unclassified", .)))

write.csv(dads.sig.dt, file="Results/ITS/deseq_trt_dads.csv",row.names = TRUE)
# Now, let's create the plot
deseqplot2 <- ggplot(dads.sig.dt, aes(y = reorder(ASV, log2FoldChange), x = log2FoldChange, fill = Phylum)) +
  geom_vline(xintercept = 0.0, color = "gray", size = 0.5, linetype = "dashed") +
  geom_bar(stat = "identity", position = "dodge", width = 0.8) +  
  scale_fill_viridis(discrete = TRUE, option = "D") +  # Color-blind friendly palette
  #scale_x_continuous(breaks = pretty_breaks(n = 10)) +  # Prettier x-axis breaks
  labs(x = "Log2 Fold Change", 
       y = "ASV",
       title = "Differential Abundance Analysis: Treatment")
  theme_minimal() +
  theme(
    axis.text.y = element_text(size = 8),
    axis.title = element_text(face = "bold"),
    legend.position = "right",
    legend.title = element_text(face = "bold"),
    plot.title = element_text(size = 16, face = "bold"),
    plot.subtitle = element_text(size = 12),
    panel.grid.major.y = element_blank(),
    panel.grid.minor = element_blank()
  )

# Save the plot
ggsave("Figures/ITS/Deseq_treatment_barplot_dads_ASV.pdf", deseqplot2, width=40, height=40)

# Display the plot
print(deseqplot2)
```

### Let's make a quick summary table of ASVs that are differentially abundant in both the Wilcoxon rank sum and DESEQ analyses-treatment
```{r, diff abundance merged analyses-trt}
# Merge the tables based on ASV
merged_table <- inner_join(dads.sig.dt, dads.wilcox.dt, by = "ASV")

# Select and rename columns in one step
final_table <- merged_table %>%
  mutate(
    deseq_pvalue = pvalue,
    deseq_padj = padj,
    wilcox_pvalue = P_Value,
    wilcox_padj = Adjusted_P_Value,
    Kingdom = Kingdom.x,
    Phylum = Phylum.x,
    Class = Class.x,
    Order = Order.x,
    Family = Family.x,
    Genus = Genus.x,
    Species = Species.x
  ) %>%
  select(
    ASV, log2FoldChange, deseq_pvalue, deseq_padj, wilcox_pvalue, wilcox_padj, 
    Kingdom, Phylum, Class, Order, Family, Genus, Species
  )


# Save the final table as CSV
write.csv(final_table, "Results/ITS/merged_deseq_trt_wilcoxon_results.csv", row.names = FALSE)
print(paste("Number of ASVs in the final table:", nrow(final_table)))
```

### Let's make a plot of the merged results-treatment
```{r plot of merged deseq and wilcoxon rank sum analyses-treatment}
merged_plot <- ggplot(final_table, aes(y = reorder(ASV, log2FoldChange), x = log2FoldChange, fill = Genus)) +
  geom_vline(xintercept = 0.0, color = "gray", size = 0.5, linetype = "dashed") +
  geom_bar(stat = "identity", position = "dodge", width = 0.8) +  
  scale_fill_viridis(discrete = TRUE, option = "D") +  # Color-blind friendly palette
  #scale_x_continuous(breaks = pretty_breaks(n = 10)) +  # Prettier x-axis breaks
  labs(x = "Log2 Fold Change", 
       y = "ASV",
       title = "Merged DESEQ2 and Wilcoxon Rank Sum Analyses-Treatment")
  theme_minimal() +
  theme(
    axis.text.y = element_text(size = 8),
    axis.title = element_text(face = "bold"),
    legend.position = "right",
    legend.title = element_text(face = "bold"),
    plot.title = element_text(size = 16, face = "bold"),
    plot.subtitle = element_text(size = 12),
    panel.grid.major.y = element_blank(),
    panel.grid.minor = element_blank()
  )

# Save the plot
ggsave("Figures/ITS/Deseq_merged_trt_barplot_dads_ASV.pdf", merged_plot, height=30, width=40)

# Display the plot
print(merged_plot)
```

### Conduct Wilcoxon-rank sum tests on the dads data set-soil
```{r wilcoxon-tests-soil}
set.seed(1)
wilcox.tests <- lapply(dads.asv.ids, function(asv) {
  frm <- paste(asv, "~ Soil") # this combined with the as.formula() function below makes it possible to generate formulas in R programmatically
  return(
    wilcox.test(as.formula(frm), data = dads.dt)
  )
})

dads.raw.pvals <- sapply(wilcox.tests, function(mod) {return(mod$p.value)})

dads.fdr.pvals <- p.adjust(dads.raw.pvals, method = "fdr")

paste("Significant associations:", length(dads.fdr.pvals[dads.fdr.pvals <= 0.05])) %>%
  cat.n()

dads.raw.pvals <- sapply(wilcox.tests, function(mod) mod$p.value)
dads.fdr.pvals <- p.adjust(dads.raw.pvals, method = "fdr")

# Create data table of significant results
dads.wilcox.dt <- data.table(
  ASV = dads.asv.ids,
  P_Value = dads.raw.pvals,
  Adjusted_P_Value = dads.fdr.pvals
)

dads.wilcox.dt <- dads.wilcox.dt[order(Adjusted_P_Value)]

# Filter rows where Adjusted_P_Value < 0.05
dads.wilcox.dt <- dads.wilcox.dt[Adjusted_P_Value < 0.05]

# Optionally, merge taxonomic information if available
tax_table <- as.data.frame(tax_table(dads_5asv.ps))
tax_table$ASV_id <- rownames(tax_table)
dads.wilcox.dt <- merge(dads.wilcox.dt, tax_table, by.x = "ASV", by.y = "ASV_id", all.x = TRUE)

# Export to CSV
write.csv(dads.wilcox.dt, file = "Results/ITS/wilcox_soil_dads.csv", row.names = FALSE)
```

### Let's now do a DESEQ2 analysis-soil
```{r deseq2-soil}
set.seed(1)
dads.deseq <- phyloseq_to_deseq2(dads_5asv.ps, ~ Soil) # this function does all the heavy lifting for us to turn a phyloseq object into a DESeq2 object

#This was required so DESEQ doesn't error out (there are several approaches, but the deseq developer recommended this most recently, if going from phyloseq to deseq option)
dads.dds <- estimateSizeFactors(dads.deseq, type = 'poscounts')
dads.dds = DESeq(dads.dds)
dads.res.dt <- results(dads.dds) %>% as.data.table(keep.rownames = "ASV")

dads.sig.dt <- dads.res.dt[padj <= 0.05]

dads.sig.dt = cbind(as(dads.sig.dt, "data.frame"), as(tax_table(dads_5asv.ps)[dads.sig.dt$ASV, ], "matrix"))

dads.sig.dt <- dads.sig.dt %>%
  mutate(across(everything(), ~ifelse(is.na(.), "Unclassified", .)))

write.csv(dads.sig.dt, file="Results/ITS/deseq_soil_dads.csv",row.names = TRUE)
# Now, let's create the plot
deseqplot2 <- ggplot(dads.sig.dt, aes(y = reorder(ASV, log2FoldChange), x = log2FoldChange, fill = Genus)) +
  geom_vline(xintercept = 0.0, color = "gray", size = 0.5, linetype = "dashed") +
  geom_bar(stat = "identity", position = "dodge", width = 0.8) +  
  scale_fill_viridis(discrete = TRUE, option = "D") +  # Color-blind friendly palette
  #scale_x_continuous(breaks = pretty_breaks(n = 10)) +  # Prettier x-axis breaks
  labs(x = "Log2 Fold Change", 
       y = "ASV",
       title = "Differential Abundance Analysis: Soil")
  theme_minimal() +
  theme(
    axis.text.y = element_text(size = 8),
    axis.title = element_text(face = "bold"),
    legend.position = "right",
    legend.title = element_text(face = "bold"),
    plot.title = element_text(size = 16, face = "bold"),
    plot.subtitle = element_text(size = 12),
    panel.grid.major.y = element_blank(),
    panel.grid.minor = element_blank()
  )

# Save the plot
ggsave("Figures/ITS/Deseq_soil_barplot_dads_ASV.pdf", deseqplot2)

# Display the plot
print(deseqplot2)
```

### Let's make a quick summary table of ASVs that are differentially abundant in both the Wilcoxon rank sum and DESEQ analyses-soil
```{r, diff abundance merged analyses soil}
# Merge the tables based on ASV
merged_table <- inner_join(dads.sig.dt, dads.wilcox.dt, by = "ASV")

# Select and rename columns in one step
final_table <- merged_table %>%
  mutate(
    deseq_pvalue = pvalue,
    deseq_padj = padj,
    wilcox_pvalue = P_Value,
    wilcox_padj = Adjusted_P_Value,
    Kingdom = Kingdom.x,
    Phylum = Phylum.x,
    Class = Class.x,
    Order = Order.x,
    Family = Family.x,
    Genus = Genus.x,
    Species = Species.x
  ) %>%
  select(
    ASV, log2FoldChange, deseq_pvalue, deseq_padj, wilcox_pvalue, wilcox_padj, 
    Kingdom, Phylum, Class, Order, Family, Genus, Species
  )


# Save the final table as CSV
write.csv(final_table, "Results/ITS/merged_deseq_soil_wilcoxon_results.csv", row.names = FALSE)
print(paste("Number of ASVs in the final table:", nrow(final_table)))
```

### Let's make a plot of the merged results-soil
```{r plot of merged deseq and wilcoxon analyses-soil}
merged_plot <- ggplot(final_table, aes(y = reorder(ASV, log2FoldChange), x = log2FoldChange, fill = Genus)) +
  geom_vline(xintercept = 0.0, color = "gray", size = 0.5, linetype = "dashed") +
  geom_bar(stat = "identity", position = "dodge", width = 0.8) +  
  scale_fill_viridis(discrete = TRUE, option = "D") +  # Color-blind friendly palette
  #scale_x_continuous(breaks = pretty_breaks(n = 10)) +  # Prettier x-axis breaks
  labs(x = "Log2 Fold Change", 
       y = "ASV",
       title = "Merged DESEQ2 and Wilcoxon Rank Sum Analyses-Treatment")
  theme_minimal() +
  theme(
    axis.text.y = element_text(size = 8),
    axis.title = element_text(face = "bold"),
    legend.position = "right",
    legend.title = element_text(face = "bold"),
    plot.title = element_text(size = 16, face = "bold"),
    plot.subtitle = element_text(size = 12),
    panel.grid.major.y = element_blank(),
    panel.grid.minor = element_blank()
  )

# Save the plot
ggsave("Figures/ITS/Deseq_merged_soil_barplot_dads_ASV.pdf", merged_plot)

# Display the plot
print(merged_plot)
```

### Conduct Wilcoxon-rank sum tests on the dads data set-incubation
```{r wilcoxon-tests incubation}
set.seed(1)
wilcox.tests <- lapply(dads.asv.ids, function(asv) {
  frm <- paste(asv, "~ Incubation") # this combined with the as.formula() function below makes it possible to generate formulas in R programmatically
  return(
    wilcox.test(as.formula(frm), data = dads.dt)
  )
})

dads.raw.pvals <- sapply(wilcox.tests, function(mod) {return(mod$p.value)})

dads.fdr.pvals <- p.adjust(dads.raw.pvals, method = "fdr")

paste("Significant associations:", length(dads.fdr.pvals[dads.fdr.pvals <= 0.05])) %>%
  cat.n()

dads.raw.pvals <- sapply(wilcox.tests, function(mod) mod$p.value)
dads.fdr.pvals <- p.adjust(dads.raw.pvals, method = "fdr")

# Create data table of significant results
dads.wilcox.dt <- data.table(
  ASV = dads.asv.ids,
  P_Value = dads.raw.pvals,
  Adjusted_P_Value = dads.fdr.pvals
)

dads.wilcox.dt <- dads.wilcox.dt[order(Adjusted_P_Value)]

# Filter rows where Adjusted_P_Value < 0.05
dads.wilcox.dt <- dads.wilcox.dt[Adjusted_P_Value < 0.05]

# Optionally, merge taxonomic information if available
tax_table <- as.data.frame(tax_table(dads_5asv.ps))
tax_table$ASV_id <- rownames(tax_table)
dads.wilcox.dt <- merge(dads.wilcox.dt, tax_table, by.x = "ASV", by.y = "ASV_id", all.x = TRUE)

# Export to CSV
write.csv(dads.wilcox.dt, file = "Results/ITS/wilcox_incubation_dads.csv", row.names = FALSE)
```

### Let's now do a DESEQ2 analysis-incubation-no significant results.
```{r deseq2-incubation}
set.seed(1)
dads.deseq <- phyloseq_to_deseq2(dads_5asv.ps, ~ Incubation) # this function does all the heavy lifting for us to turn a phyloseq object into a DESeq2 object

#This was required so DESEQ doesn't error out (there are several approaches, but the deseq developer recommended this most recently, if going from phyloseq to deseq option)
dads.dds <- estimateSizeFactors(dads.deseq, type = 'poscounts')
dads.dds = DESeq(dads.dds)
dads.res.dt <- results(dads.dds) %>% as.data.table(keep.rownames = "ASV")

#dads.sig.dt <- dads.res.dt[padj <= 0.05]

#dads.sig.dt = cbind(as(dads.sig.dt, "data.frame"), #as(tax_table(dads_5asv.ps)[dads.sig.dt$ASV, ], "matrix"))
# 
#dads.sig.dt <- dads.sig.dt %>%
#mutate(across(everything(), ~ifelse(is.na(.), "Unclassified", .)))
# 
#write.csv(dads.sig.dt, # #file="Results/ITS/deseq_incubation_dads.csv",row.names = TRUE)
# 
# # Now, let's create the plot
# deseqplot2 <- ggplot(dads.sig.dt, aes(y = reorder(ASV, log2FoldChange), x = log2FoldChange, fill = Genus)) +
#   geom_vline(xintercept = 0.0, color = "gray", size = 0.5, linetype = "dashed") +
#   geom_bar(stat = "identity", position = "dodge", width = 0.8) +  
#   scale_fill_viridis(discrete = TRUE, option = "D") +  # Color-blind friendly palette
#   #scale_x_continuous(breaks = pretty_breaks(n = 10)) +  # Prettier x-axis breaks
#   labs(x = "Log2 Fold Change", 
#        y = "ASV",
#        title = "Differential Abundance Analysis: Incubation")
#   theme_minimal() +
#   theme(
#     axis.text.y = element_text(size = 8),
#     axis.title = element_text(face = "bold"),
#     legend.position = "right",
#     legend.title = element_text(face = "bold"),
#     plot.title = element_text(size = 16, face = "bold"),
#     plot.subtitle = element_text(size = 12),
#     panel.grid.major.y = element_blank(),
#     panel.grid.minor = element_blank()
#   )
# 
# # Save the plot
# ggsave("Figures/ITS/Deseq_incubation_barplot_dads_ASV.pdf", deseqplot2)
# 
# # Display the plot
# print(deseqplot2)
```

### Let's make a quick summary table of ASVs that are differentially abundant in both the Wilcoxon rank sum and DESEQ analyses-incubation
```{r, diff abundance merged analyses incubation}
# # Merge the tables based on ASV
# merged_table <- inner_join(dads.sig.dt, dads.wilcox.dt, by = "ASV")
# 
# # Select and rename columns in one step
# final_table <- merged_table %>%
#   mutate(
#     deseq_pvalue = pvalue,
#     deseq_padj = padj,
#     wilcox_pvalue = P_Value,
#     wilcox_padj = Adjusted_P_Value,
#     Kingdom = Kingdom.x,
#     Phylum = Phylum.x,
#     Class = Class.x,
#     Order = Order.x,
#     Family = Family.x,
#     Genus = Genus.x,
#     Species = Species.x
#   ) %>%
#   select(
#     ASV, log2FoldChange, deseq_pvalue, deseq_padj, wilcox_pvalue, wilcox_padj, 
#     Kingdom, Phylum, Class, Order, Family, Genus, Species
#   )
# 
# # Save the final table as CSV
# write.csv(final_table, "Results/ITS/merged_deseq_incubation_wilcoxon_results.csv", row.names = FALSE)
# print(paste("Number of ASVs in the final table:", nrow(final_table)))
```

### Let's make a plot of the merged results-incubation

In this case there is nothing to merge
```{r plot of merged deseq and wilcoxon analyses-incubation}
#There is nothing to merge
# merged_plot <- ggplot(final_table, aes(y = reorder(ASV, log2FoldChange), x = log2FoldChange, fill = Genus)) +
#   geom_vline(xintercept = 0.0, color = "gray", size = 0.5, linetype = "dashed") +
#   geom_bar(stat = "identity", position = "dodge", width = 0.8) +  
#   scale_fill_viridis(discrete = TRUE, option = "D") +  # Color-blind friendly palette
#   #scale_x_continuous(breaks = pretty_breaks(n = 10)) +  # Prettier x-axis breaks
#   labs(x = "Log2 Fold Change", 
#        y = "ASV",
#        title = "Merged DESEQ2 and Wilcoxon Rank Sum Analyses-Treatment")
#   theme_minimal() +
#   theme(
#     axis.text.y = element_text(size = 8),
#     axis.title = element_text(face = "bold"),
#     legend.position = "right",
#     legend.title = element_text(face = "bold"),
#     plot.title = element_text(size = 16, face = "bold"),
#     plot.subtitle = element_text(size = 12),
#     panel.grid.major.y = element_blank(),
#     panel.grid.minor = element_blank()
#   )
# 
# # Save the plot
# ggsave("Figures/ITS/Deseq_merged_incubation_barplot_dads_ASV.pdf", merged_plot)
# 
# # Display the plot
# print(merged_plot)
```

### For nicer visuals, let's make heat trees (specifically the comparison between untreated and treated samples), and we'll start with Wilcoxon rank sum-treatment

### Convert taxmap to phyloseq obj
```{r, convert to phyloseq obj}
remove_boot_columns <- function(x) {
  x <- x[, !grepl("boot", colnames(x))]
  return(x)
}

adjust_species_names <- function(x) {
  x$Species <- sub(".*_", "", x$Species)
  return(x)
}

dads_5asv_tax <- tax_table(dads_5asv.ps) %>%
  as.data.frame() %>%
  remove_boot_columns() %>%
  adjust_species_names %>%
  as.matrix() %>%
  tax_table()
  
tax_table(dads_5asv.ps) <- dads_5asv_tax

obj_ITS<-parse_phyloseq(dads_5asv.ps)

obj_ITS$data$tax_abund <- calc_taxon_abund(obj_ITS, data = "otu_table", cols = obj_ITS$data$sample_data$sample_id)

obj_ITS$data$abund_prop <- calc_obs_props(obj_ITS, data = "otu_table", cols = obj_ITS$data$sample_data$sample_id)

obj_ITS$data$tax_prop <- calc_taxon_abund(obj_ITS, data = "abund_prop", cols = obj_ITS$data$sample_data$sample_id)

obj_ITS$data$tax_prop$all_samples <- rowMeans(obj_ITS$data$tax_prop[,obj_ITS$data$sample_data$sample_id])
```

### We'll then do DESEQ analysis on untreated vs. untreated samples
```{r, diff abundance trees-deseq}
set.seed(1)
obj_ITS$data$diff_table_deseq <- metacoder::calc_diff_abund_deseq2(obj_ITS,
                                      data = "tax_abund",
                                      cols = obj_ITS$data$sample_data$sample_id, # What columns of sample data to
                                      groups = obj_ITS$data$sample_data$Treatment) # What category each sample is assigned to

obj_ITS$data$diff_table_deseq$log2FoldChange[obj_ITS$data$diff_table_deseq$padj > 0.05] <- 0

range(obj_ITS$data$diff_table_deseq$padj, finite = TRUE) 
range(obj_ITS$data$diff_table_deseq$log2FoldChange, finite = TRUE) 

obj_ITS<- metacoder::filter_taxa(obj_ITS, taxon_names != "None", reassign_obs = FALSE)
obj_ITS <- metacoder::filter_taxa(obj_ITS, taxon_names != "", reassign_obs = FALSE)
obj_ITS <- metacoder::filter_taxa(obj_ITS, taxon_names != "NA", reassign_obs = FALSE)

obj_ITS %>%
  metacoder::filter_taxa(taxon_ranks == "Species", supertaxa = TRUE, reassign_obs = FALSE) %>%
  remove_redundant_names() %>%
  heat_tree_matrix(data = "diff_table_deseq",
                   node_label = taxon_names,
                   node_size = n_obs, # number of OTUs
                   node_color = log2FoldChange, # difference between groups
                   node_color_trans = "linear",
                   node_color_interval = c(-12, 12), # symmetric interval
                   edge_color_interval = c(-12, 12), # symmetric interval
                   node_color_range = diverging_palette(), # diverging colors
                   node_size_axis_label = "ASV count",
                   node_color_axis_label = "Log 2 Fold Change",
                   layout = "da", initial_layout = "re",
                   key_size = 0.4,
                   output_file = "Figures/ITS/diff_abundance_heattree_deseq_dads_ITS.pdf") # The layout algorithm that initializes node locations
```

### Software used
```{r software used}
sessioninfo::session_info()
```
